% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared.R
\name{chi_squared}
\alias{chi_squared}
\title{Chi-squared test}
\usage{
chi_squared(data, alpha = 0.05)
}
\arguments{
\item{data}{A data frame with samples as columns and bins as rows.}

\item{alpha}{Significance level (default 0.05).}
}
\value{
A data frame with one row, containing the output of the Chi-squared test.
}
\description{
Chi-squared test
}
\examples{
example_data <- data.frame(Sample1 = c(10,20,30),
                           Sample2 = c(10,12,38),
                           Sample3 = c(10,22,28))
chi_squared(data = example_data, alpha = 0.05)

}
