% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_chi_squared.R
\name{pairwise_chi_squared}
\alias{pairwise_chi_squared}
\title{Pairwise Chi-squared test}
\usage{
pairwise_chi_squared(
  data,
  comparisons = "all",
  alpha = 0.05,
  adjust = c("Bonferroni", "BH", "none")
)
}
\arguments{
\item{data}{A data frame with samples as columns and bins as rows.}

\item{comparisons}{A list containing the comparisons to perform, in the format c("Sample1", "Sample2"). Default is "all", which performs all pairwise comparisons.}

\item{alpha}{Significance level (default 0.05).}

\item{adjust}{Multiple testing correction method. Must be one of "Bonferroni" (default), "BH" or "none" (not recommended).}
}
\value{
A data frame with one row per comparison, containing the output of the Chi-squared test.
}
\description{
Pairwise Chi-squared test
}
\examples{
example_data <- data.frame(Sample1 = c(10,20,30),
                           Sample2 = c(10,12,38),
                           Sample3 = c(10,22,28))
pairwise_chi_squared(data = example_data,
                     comparisons = list(c("Sample1", "Sample2"),
                                        c("Sample1", "Sample3")),
                     alpha = 0.05,
                     adjust = "Bonferroni")

}
